% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_templin}
\alias{base_templin}
\title{Templin baseline data.}
\format{
A tibble with 29363 rows and 60 variables:
}
\source{
ADF&G GCL database LOKI
}
\usage{
base_templin
}
\description{
An example of genotype data for preparing msgsi input data
in \emph{rubias} format (but prep_msgsi_data() can also take GCL format)
for making road-scale baseline
}
\keyword{datasets}
