% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_yukon}
\alias{base_yukon}
\title{Yukon baseline data.}
\format{
A tibble with 5435 rows and 358 variables:
}
\source{
ADF&G GCL database LOKI
}
\usage{
base_yukon
}
\description{
An example of genotype data for preparing msgsi input data
in \emph{rubias} format (but prep_msgsi_data() can also take GCL format)
for making regional, finer-scale baseline
}
\keyword{datasets}
