% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{ggtr_plt}
\alias{ggtr_plt}
\title{Plot MCMC trace}
\usage{
ggtr_plt(obj, nburn = 0, thin = 1, name_order = NULL)
}
\arguments{
\item{obj}{Trace from the model output.}

\item{nburn}{Number of burn-in you set up when you ran the model. Default is 0 if you didn't save the burn-ins (keep_burn = FALSE).}

\item{thin}{Number of thinning you set up when you ran the model. Default is 1 (no thinning).}

\item{name_order}{Arrange the reporting groups as you wish. Leave it empty if you want to accept the default.}
}
\value{
Trace plot in ggplot
}
\description{
Plot MCMC trace
}
\examples{
nu_order <- c("Kenai_Pen", "UCI_Northwest", "Susitna_Mainstem", "Knik_Turnagain", "Deshka", "Yentna")
ggtr_plt(obj = msgsi_out$trace, nburn = 2500, thin = 5, name_order = nu_order)
}
