% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_assign.R
\name{indiv_assign}
\alias{indiv_assign}
\title{Individual assignment summary}
\usage{
indiv_assign(mdl_out, mdl_dat, show_t2_grps = TRUE)
}
\arguments{
\item{mdl_out}{Model output object name.}

\item{mdl_dat}{Input data object name.}

\item{show_t2_grps}{Set it to be \code{FALSE} if you want the proportions of broad-scale groups to be combined. Default = \code{TRUE}.}
}
\value{
Individual assignment summary
}
\description{
Individual assignment summary
}
\examples{
# set up input data and run multistage model
msgsi_dat <-
  prep_msgsi_data(mixture_data = mix,
  baseline1_data = base_templin, baseline2_data = base_yukon,
  pop1_info = templin_pops211, pop2_info = yukon_pops50, sub_group = 3:5)

# run model
msgsi_out <- msgsi_mdl(msgsi_dat, nreps = 25, nburn = 15, thin = 1, nchains = 1)

# individual assignment summary
ind_iden <- indiv_assign(msgsi_out, msgsi_dat)

}
