% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mix}
\alias{mix}
\title{Mixture data.}
\format{
A tibble with 150 rows and 358 variables:
}
\source{
ADF&G GCL database LOKI
}
\usage{
mix
}
\description{
An example of genotype data for preparing msgsi input data
in \emph{rubias} format (but prep_msgsi_data() can also take GCL format)
for making mixed population data
This is a made up mixture sample, not real
}
\keyword{datasets}
