% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgsi_mdl.R
\name{msgsi_mdl}
\alias{msgsi_mdl}
\title{Run multistage GSI model}
\usage{
msgsi_mdl(
  dat_in,
  nreps,
  nburn,
  thin,
  nchains,
  nadapt = 0,
  keep_burn = FALSE,
  cond_gsi = TRUE,
  out_path = NULL,
  seed = NULL
)
}
\arguments{
\item{dat_in}{Name of the input data.}

\item{nreps}{Total number of iterations (includes burn-ins).}

\item{nburn}{Number of warm-up runs.}

\item{thin}{Frequency to thin the output.}

\item{nchains}{Number of independent MCMC processes.}

\item{nadapt}{Number of adaptation run (default is 0). Only available when running model in fully Bayesian mode.}

\item{keep_burn}{To save the burn-ins or not (default is FALSE).}

\item{cond_gsi}{To run the model in conditional GSI mode (default is TRUE).}

\item{out_path}{File path to save the output. Leave it empty is you don't want to save the output. Default is NULL (no saving).}

\item{seed}{Random seed for reproducibility. Default is NULL (no random seed).}
}
\value{
A list contains reporting group proportion summary and trace for tier one (summ_t1, trace_t1), tier two (summ_t2, trace_t2) and combined two tiers (summ_comb, trace_comb), and record of individual assignment during first tier for each individual (idens).
}
\description{
Run multistage GSI model
}
\examples{
msgsi_out <- msgsi_mdl(msgsi_data, nreps = 25, nburn = 15, thin = 1, nchains = 5, nadapt = 0, keep_burn = FALSE, cond_gsi = TRUE, out_path = "v:/serious_analysis_files/gsi/msgsi_output.RData", seed = 555)
}
