% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgsi_data.R
\name{prep_msgsi_data}
\alias{prep_msgsi_data}
\title{Preparing multistage GSI input data}
\usage{
prep_msgsi_data(
  mixture_data,
  baseline1_data,
  baseline2_data,
  pop1_info,
  pop2_info,
  sub_group,
  file_path = NULL,
  loci1 = NULL,
  loci2 = NULL
)
}
\arguments{
\item{mixture_data}{Individual fish with loci for both tier one and tier two. Mixture data in GCL or rubias format.}

\item{baseline1_data}{Tier one baseline data in GCL or rubias format.}

\item{baseline2_data}{Tier two baseline data in GCL or rubias format.}

\item{pop1_info}{Population information for tier one. A tibble with columns collection (collection names), repunit (reporting unit names), grpvec (group numbers), origin (wild/hatchery).}

\item{pop2_info}{Population information for tier two. A tibble with columns collection (collection names), repunit (reporting unit names), grpvec (group numbers).}

\item{sub_group}{Group numbers for groups of interest. Group id numbers in tier one that identify groups in tier two.}

\item{file_path}{Where you want to save a copy of input data. Leave it blank if you don't want to save a copy.}

\item{loci1}{Optional. Provide loci (for tier one) as a fail-safe check.}

\item{loci2}{Optional. Provide loci (for tier two) as a fail-safe check.}
}
\value{
A list objects as the input data for msgsi_mdl()
}
\description{
Preparing multistage GSI input data
}
\examples{
msgsi_dat <- prep_msgsi_data(mixture_data = mix.gcl, baseline1_data = base_gcl, pop1_info = uci_pops55, pop2_info = ucigtseq_pops66 \%>\% filter(grpvec != 6), sub_group = c(1,2,4:6), file_path = "uci_chinook/rabies/msgsi_dat_uci.RData", loci1 = loci34, loci2 = loci_gtseq)
}
