% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templin_pops211}
\alias{templin_pops211}
\title{Templin population infomation.}
\format{
A tibble with 211 rows and 3 variables: \code{collection}, \code{repunit},
\code{grpvec}
}
\usage{
templin_pops211
}
\description{
A tibble that contains collection name, reporting unit, group number
for each of the 211 populations in the Templin baseline
}
\keyword{datasets}
