% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{tr_plot}
\alias{tr_plot}
\title{Plot MCMC trace}
\usage{
tr_plot(mdl_out, trace_obj, pop_info = NULL)
}
\arguments{
\item{mdl_out}{Model output object name.}

\item{trace_obj}{Trace from the model output.}

\item{pop_info}{Population information. A tibble with columns
collection (collection names), repunit (reporting unit names),
and grpvec (group numbers).}
}
\value{
Trace plot in ggplot
}
\description{
Make trace plot for visualizing the mixing of MCMC chains. Group information (as in \code{groups}, \code{p2_groups}, or \code{comb_groups}) should be provided for the trace history that you want to plot or the trace plots will be shown as collections. See vignette for details.
}
\examples{
# set up input data and run multistage model
msgsi_dat <-
  prep_msgsi_data(mixture_data = mix,
  baseline1_data = base_templin, baseline2_data = base_yukon,
  pop1_info = templin_pops211, pop2_info = yukon_pops50, sub_group = 3:5)

msgsi_out <- msgsi_mdl(msgsi_dat, nreps = 25, nburn = 15, thin = 1, nchains = 1)

# trace plot
tr_plot(mdl_out = msgsi_out, trace_obj = "trace_comb", pop_info = msgsi_out$comb_groups)
}
