% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yukon_pops50}
\alias{yukon_pops50}
\title{Yukon population infomation.}
\format{
A tibble with 50 rows and 3 variables: \code{collection}, \code{repunit},
\code{grpvec}
}
\usage{
yukon_pops50
}
\description{
A tibble that contains collection name, reporting unit, group number
for each of the 50 populations in the Yukon River baseline
}
\keyword{datasets}
