% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgsi_data.R
\name{prep_msgsi_data}
\alias{prep_msgsi_data}
\title{Preparing multistage GSI input data}
\usage{
prep_msgsi_data(
  mixture_data,
  baseline1_data,
  baseline2_data,
  pop1_info,
  pop2_info,
  sub_group,
  file_path = NULL,
  loci1 = NULL,
  loci2 = NULL
)
}
\arguments{
\item{mixture_data}{Individual fish with loci for both tier 1 and tier 2.
Mixture data in GCL or \emph{rubias} format.}

\item{baseline1_data}{Tier 1 baseline data in GCL or \emph{rubias} format.}

\item{baseline2_data}{Tier 2 baseline data in GCL or \emph{rubias} format.}

\item{pop1_info}{Population information for tier 1. A tibble with columns
collection (collection names), repunit (reporting unit names),
grpvec (group numbers), origin (wild/hatchery).}

\item{pop2_info}{Population information for tier 2. A tibble with columns
collection (collection names), repunit (reporting unit names),
grpvec (group numbers).}

\item{sub_group}{Group numbers for groups of interest. Group id numbers in tier
1 that identify groups in tier 2.}

\item{file_path}{Where you want to save a copy of input data. Leave it empty if
you don't want to save a copy.}

\item{loci1}{Optional. Provide loci (for tier 1) as a fail-safe check.}

\item{loci2}{Optional. Provide loci (for tier 2) as a fail-safe check.}
}
\value{
A list objects as the input data for msgsi_mdl()
}
\description{
Preparing multistage GSI input data
}
\examples{
msgsi_dat <-
  prep_msgsi_data(mixture_data = mix,
  baseline1_data = base_templin, baseline2_data = base_yukon,
  pop1_info = templin_pops211, pop2_info = yukon_pops50, sub_group = 3:5)
}
