% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stra_est.R
\name{stratified_estimator_msgsi}
\alias{stratified_estimator_msgsi}
\title{Stratified estimator for Ms.GSI}
\usage{
stratified_estimator_msgsi(
  mdl_out = NULL,
  path = NULL,
  mixvec,
  catchvec,
  cv = NULL
)
}
\arguments{
\item{mdl_out}{Optional. Model output object name.}

\item{path}{Where to find output from each mixture as a folder.}

\item{mixvec}{Character vector of mixture sillies that are used to locate the folders where output .csv files lives, if \code{mdl_out} is not provided.}

\item{catchvec}{Numeric vector of harvest for each mixture/stratum, must be in the same order as \code{mixvec}.}

\item{cv}{Numeric vector of harvest estimate coefficients of variation for each stratum, must be the same order as \code{mixvec}.}
}
\value{
A tibble of proportions and harvest numbers by reporting group for combined mixtures/strata.
}
\description{
Stratified estimator for Ms.GSI
}
\examples{
stratified_estimator_msgsi(path = "test", mixvec = c("m1", "m2", "m3"), catchvec = c(4500, 5000, 3000), cv = c(0.5, 0.1 ,0.3))

}
